Removed a few stray cases of "uint"
After my cleanup of chpltypes.h the other day, our code still
contained a few stray cases of uint that went undetected on
our linux platforms for reasons I'm not sure of -- perhaps
our stdint.h/inttypes.h files define that alias?  In any case,
they tripped up the MTA compiler, and I don't think we want
to use them, so I removed them.



git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@15180 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ compiler/include/vec.h
@@ -375,7 +375,7 @@ template <class C, int S> C *
 Vec<C,S>::set_add_internal(C c) {
   int j, k;
   if (n) {
-    uint h = (uint)(uintptr_t)c;
+    uintptr_t h = (uintptr_t)c;
     h = h % n;
     for (k = h, j = 0;
          k < n && j < SET_MAX_PROBE;
@@ -400,7 +400,7 @@ template <class C, int S> C *
 Vec<C,S>::set_in_internal(C c) {
   int j, k;
   if (n) {
-    uint h = (uint)(uintptr_t)c;
+    uintptr_t h = (uintptr_t)c;
     h = h % n;
     for (k = h, j = 0;
          k < n && j < SET_MAX_PROBE;
